Create 752.-OpenTheLock.js
To solve this problem, you can use a breadth-first search (BFS) algorithm to explore all possible combinations of the lock's wheels. You'll start from the initial state "0000" and move towards the target state while avoiding the dead ends. Here's a JavaScript implementation of the openLock function:

This function maintains a deadSet to store dead ends and a visited set to keep track of visited states. It iteratively explores the possible states using a BFS approach until it reaches the target or exhausts all possibilities. Finally, it returns the minimum number of moves required to reach the target or -1 if it's not possible.