Create ♦♦2370-LongestIdealSubsequence.js
The JavaScript code defines a function longestIdealString that takes a string s and a number k as input and returns the length of the longest ideal string. The function uses dynamic programming to solve the problem.


It first initializes an array dp of size n (the length of the string) with all elements set to 0 and an array last_occurrence of size 26 (the number of lowercase English letters) with all elements set to -1.


Then it iterates over each character in the string. For each character, it sets dp[i] to 1 and then checks all previous characters that can be included in the ideal string (those with absolute difference in alphabet order less than or equal to k). If such a character is found, it updates dp[i] to the maximum of dp[i] and dp[last_occurrence[j]] + 1.


Finally, it updates last_occurrence[s.charCodeAt(i) - 'a'.charCodeAt(0)] to i. The function returns the maximum value in dp, which is the length of the longest ideal string.