Create 1237-FindPositiveIntegerSolutionFor_a _GivenEquation.js
This function iterates through positive integer pairs (x, y) starting from (1, 1000) and moves towards (1000, 1) while checking if customfunction.f(x, y) equals z. If it does, it adds the pair to the result array. If the value is less than z, it increments x; if it's greater, it decrements y. This ensures that the function explores all possible pairs efficiently.