Create 834-SumOfDistancesInTree.js
The solution to this problem involves using a depth-first search (DFS) approach. The idea is to first calculate the sum of distances in the subtree of every node and the number of nodes in the subtree of every node. Then, using this information, we can calculate the sum of distances in the tree for every node.

Explanation

This function first constructs the adjacency list representation of the graph. The adjacency list is a common way to represent a graph in which each node i is associated with a list of nodes that i is connected to.


Then it performs a depth-first search to calculate the sum of distances in the subtree of every node and the number of nodes in the subtree of every node. The DFS function traverses the graph in a depth-first manner, visiting each node and its descendants before backtracking.


Finally, it performs another depth-first search to calculate the sum of distances in the tree for every node. This is done by subtracting the sum of distances in the subtree of a node and the number of nodes in the subtree from the sum of distances in the tree and the total number of nodes, respectively.


The result is an array where the i-th element is the sum of distances between the i-th node in the tree and all other nodes.