Create 260-Single_Number_III.js
To solve the problem of finding two unique numbers in an array where all other numbers appear exactly twice, we can utilize bitwise operations to achieve a linear runtime complexity and constant space usage. Here's how we can approach this problem:

XOR all the numbers: This will help us in isolating the XOR of the two unique numbers since the XOR of two identical numbers is zero and the XOR operation is both commutative and associative.

Find a set bit: Identify a bit that is set (i.e., 1) in the XOR result. This bit will be different between the two unique numbers, allowing us to use it to partition the array into two groups.

Partition the array and XOR within groups: By partitioning the array based on whether the identified bit is set or not, we can XOR the numbers within each group to isolate the unique numbers.

Explanation:
XOR all numbers: We get the combined XOR of the two unique numbers because all duplicate numbers cancel out to zero.

Find a set bit: The result of XOR will have at least one bit set. We find the rightmost set bit to use as a mask.

Partition and XOR within groups: Using the mask, we partition the numbers into two groups: one group where this bit is set and another where it is not. Each group will have one unique number, and all other numbers in each group will cancel out when XORed together.

This approach ensures we achieve the required linear runtime complexity and constant space usage.